+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 4.30-SNAPSHOT
 ADDED:
  
 - new factory methods for LinkedHashSet and InsertionOrderSet created
+- new hashcode method for collections that contains array objects
 
 CHANGED: